+++ docs/changes.rst
@@ -20,6 +20,7 @@ Released on ? 2019.
 * Added check for Kashida letters.
 * Added option to squash commits based on authors.
 * Improved support for xlsx file format.
+* Compatibility with tesseract 4.0.
 
 weblate 3.4
 -----------